FishLight: Change the night brightness
Changing the night brightness to be much less bright.  The blue LEDs are
ridiculously intense.  So I am toning them down a bit.  Also, I am
adding a little bit of the Warm White because I think it makes it look
much more natural to not just be blue.  May change this to Cool White in
the future depending on how I like the warm white look.