FishLight: Add new RTC commands
Add functions RtcWrite, RtcRead, and getClkStatus to the time file.
These command can be used to write and read a specific register in the
RTC, and check the clk data pin.  The Write and Read functions need to
be updated to Write and Read more than one byte at a time because doing
it 1 byte at a time is slow and has lots of overhead.