FishLight: Updating cursors and main screen
Updated cursors to take advantage of the LCD #defines for its size.
This will make it easy to update the size if I want to get a new one.
Also added '<' and '>' to the main screen so I can add menu
functionality.  Moved the other display parts accordingly (for time,
brightness, etc so it all aligns).  I will need to move the select area
for changing the time using the buttons next to align to the new area.
Maybe this should be defined somewhere global.