FishLight: Fix assorted errors
Fixing some errors that caused the fish light to freeze
or not compile.  ALso added some LCD debug statements.
Also changed the location of the Serial.begin statement.

This appears to be stable so far.  I was having the
board freeze on me randomly but it appears to be fixed.
There appears to be some serious issue with the 'debug'
commands because if I compile them, the board dies.
I am not sure what this is about yet, possibly too big
of an 'if' statement?  To be resolved later since I
don't care besides the faster clock counting in debug
brightness function.  Probably just tie this to a
serial command.

Signed-off-by: Michael Loptien <loptienm@gmail.com>