FishLight: Change night max brightness lower
Change the max brightness to be a fixed lower value of only .01 *
MaxBrightPWM.  The night LEDs are really bright so making them about as
dim as possible in order to be a better night time brightness.  Also,
the mix of the blue and the warm white LEDs makes a nice purple color in
areas where the have lots of room to mix.

Also added in nicer file header comments.

Also changed the location of the time hot spot when using the buttons.
This should probably be a macro definition for easy updates.