FishLight: Add prelim brightness menu functions
Adding a preliminary function to set the brightness of the LEDs by using
the buttons.  My idea is to setup a desired number of peaks (0-2 so far)
where the brightness will be at a maximum.  0 is off always, 1 is to
slowly ramp up to a consistent brightness, and 2 is to come up, go down,
and come back up again.  See the code in fish_menus to see what they
should look like on the LCD screen.  I originally thought that 2 peaks
would be good for the night light so that I could see the fish very
easily during 2 peak times but now I'm thinking that is not as useful.
The blue LEDs are way too bright for their own good.