FishLight: Fix button functions and add debug info
Fix the button functions so that they don't go crazy and flash the LCD
really fast and update incorrectly.  The issue was trying to update too
many button flags on each loop so the getLCDblStatus function would
freak out because of its idle_flag being updated all the time and trying
to count then getting set back to 0.  Added a dbncKey (debounce key)
variable to hold the value of the buttons at the start of debouncing and
changed keyReleased default to 0 because a key was not released at time
0.  Also changed the key held short interval (time it takes before
cursor starts updating fast) to be shorter so that the cursor gets about
half way across the screen before it starts updating faster.  Also had
to change how the LCD backlight gets updated (changed its conditions on
setting idle_flag to 1).  Now watches for button presses as well as an
idle_flag update.

Also added better debug messaging (still commented out) so that the
button sequences make more sense.

Also changed the range for the down button because the hardware got
messed up somehow and it resolves to a much bigger range of analog
values (probably from the screws).

Also cleaned up an extra call to getMaxBrightness in loop().