FishLight: Merging back in button improvements
Since the Arduino GUI doesn't refresh when a file has
been updated outside of the GUI, I accidentally
removed my updates for the buttons.  I am too lazy
to go through and fix it with gits other tools so
I am just going to make a new commit to put them back
in.

Signed-off-by: Mike Loptien <loptienm@gmail.com>