FishLight: Update button time and bright setting
Updated the buttons time setting capability to be more robust and to
actually be able to abandon time changes with the Y/N 'N' option.  Also
added a custom character for the '\' since my LCD controller doesn't
seem to have it.  This is actually pretty easy to do but is really ugly
to read...  Also added the capability to actually get into the buttons
LED brightness changing functions, but they aren't fully functional yet.
But they are doing things like printing menu functions.  Just can't
actually set the brightness yet.