FishLight: Add a topbar character and LED structs
Added a new LCD character to make a top bar (like an underscore but on
top) to make the peaks menu look more appropriate and fixed the
corresponding functions to initialize that new character and print it
out.

Also added structures for the LEDs.  These structures contain all the
variables needed to manipulate the LEDs in any function.  They are not
connected to anything yet but this will help with the buttons brightness
setting and making the brightness calculator more generic.