FishLight: Add Excel file to chart LED brightness
Adding a file to chart the brightness of the LEDs throughout the day.
This is an Excel spreadsheet that contains data from the brightness
calculations for each LED.  It only prints out every 10th data point
because Excel can only handle 255 points.  But that is good enough to
get me a good approximation of the shape of the LED curves.