FishLight: Got LED structures working with scheds
Got the LED variable structures working with the LED schedule function.
The brightness seems to hold to the same pattern as it did before.
HOWEVER, the Blue LED is not working and the Warm White LED does not
have an 'off brightness' besides just off.  Maybe make a 'trough' or
'on_start_bright' and 'off_time_bright' variables?  Main problem is
getting the Blue LED to work since it wraps from ~1260 - 0 - ~240 (mins
since midnight which is about 9pm - 4am).